ai-service/entrepreneur-llm/requirements.txt
openai==1.5.0
python-dotenv==1.0.0
pymongo==4.6.1
sentence-transformers==2.2.2
langchain-community==0.0.20
numpy==1.26.4
pydantic==2.6.1
fastapi==0.109.2